Refactoring bootstrapping and adding full support for constraints & multi-column indexes (#596)
* Refactoring bootstrapping and adding full support for constraints & multi-column indexes

* Removed deleted code and add some comments

* Add comment for Constraints

* Update crates/core/src/db/datastore/locking_tx_datastore/mod.rs


* Update crates/core/src/db/datastore/locking_tx_datastore/mod.rs


* Addressing some PR comments & remove cloning

* Adding details for when fail to decode rows in replay

* Apply Kim's fix for replay bug

* Addressing some PR comments

* Merge

* Fix conflicts

* Fix conflicts

* Fix test

* Fix generation of code for csharp & typescript, now it generate FilterBy for all columns

* Fix for index generation

* Fixing C# index generation

* Merge

* Fix compilation on csharp bindings

---------

+++ crates/bindings-csharp/Runtime/RawBindings.cs
@@ -54,11 +54,11 @@ public static partial class RawBindings
     }
 
     [StructLayout(LayoutKind.Sequential)]
-    public readonly struct ColId(byte col_id)
+    public readonly struct ColId(uint col_id)
     {
-        private readonly byte col_id = col_id;
+        private readonly uint col_id = col_id;
 
-        public static explicit operator byte(ColId col_id) => col_id.col_id;
+        public static explicit operator uint(ColId col_id) => col_id.col_id;
     }
 
     [StructLayout(LayoutKind.Sequential)]
